+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/CsvimIT.java
@@ -7,7 +7,7 @@
  *
  * SPDX-FileCopyrightText: Eclipse Dirigible contributors SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.integration.tests.api.java;
+package org.eclipse.dirigible.integration.tests.ui.tests;
 
 import org.eclipse.dirigible.integration.tests.ui.tests.projects.PredefinedProjectIT;
 import org.eclipse.dirigible.integration.tests.ui.tests.projects.TestProject;
